feat(room)!: add description to rooms
Make it possible to provide description when creating a new room via REST API

+++ src/main/java/net/rooms/RoomsServer/room/RoomService.java
@@ -29,9 +29,10 @@ public class RoomService {
 	public String create(CreateRequest request, User user) {
 		if (request.title().contains("\"")) return "Invalid character: \"";
 		if (request.password().contains("\"")) return "Invalid character: \"";
+		if (request.description().contains("\"")) return "Invalid character: \"";
 
 		long roomID = roomRepository.lastID() + 1;
-		Room room = new Room(roomID, request.title(), request.isPrivate(), request.password(), user.username(), LocalDateTime.now());
+		Room room = new Room(roomID, request.title(), request.isPrivate(), request.password(), user.username(), LocalDateTime.now(), request.description());
 		if (!roomRepository.create(room)) return "Room creation failed on create";
 		if (!roomRepository.joinUser(roomID, user.username())) {
 			roomRepository.delete(roomID); // Cleans the room in case the join table could not be updated.
